+++ packages/terra-site/webpack.config.js
@@ -60,6 +60,9 @@ module.exports = {
         },
         {
           loader: 'sass-loader',
+          options: {
+            data: '$bundled-themes: "mock";',
+          },
         }],
       }),
     },